test(deletion-impact): corrections post-revue sur les contributeurs wallet et payment
- WalletDeletionImpactContributorTest : ajoute le cas PENDING+PROCESSING
  → un seul constat WALLET_REFUND_PENDING de count=2 (l'agrégation est un choix,
  pas un oubli) ; supprime lenient() devenu inutile (le stub refundRepository
  est effectivement appelé car les deux repositories sont toujours interrogés)
- PaymentDeletionImpactContributorTest : ajoute l'assertion sur count() dans
  activeEscrow_isBlocking pour qu'un constat à zéro ne passe pas inaperçu
- WalletDeletionImpactContributor : remplace le cast (int) par Math.toIntExact()

+++ src/main/java/com/yadony/api/payments/wallet/WalletDeletionImpactContributor.java
@@ -42,7 +42,7 @@ public class WalletDeletionImpactContributor implements UserDeletionImpactContri
                 .count();
         if (credited > 0) {
             findings.add(ImpactFinding.plain(
-                    ImpactSeverity.BLOCKING, "WALLET_POSITIVE_BALANCE", (int) credited));
+                    ImpactSeverity.BLOCKING, "WALLET_POSITIVE_BALANCE", Math.toIntExact(credited)));
         }
 
         int unresolved = UNRESOLVED.stream()
